fix two

+++ ChipChip613/Controllers/NhapHangsController.cs
@@ -171,7 +171,7 @@ namespace ChipChip613.Controllers
             }
         }
 
-        public JsonResult GetThanhTien(decimal donGia = 0, int soLuong = 1, decimal chiPhiKhac = 0)
+        public JsonResult GetThanhTien(decimal donGia = 0, decimal soLuong = 1, decimal chiPhiKhac = 0)
         {
 
             var thanhTien = donGia * (decimal)soLuong + chiPhiKhac;
